requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
googlesearch-python
requests
beautifulsoup4
